Refactor SA search bits into a new component.
Rename search-field.js to SearchField.js.
